+++ lib/tester/AlexaTest.ts
@@ -4,9 +4,12 @@
 
 import { RequestEnvelope } from 'ask-sdk-model';
 import { fail } from 'assert';
+import { CloudFormation as CFNClient, Lambda as LambdaClient } from 'aws-sdk';
 import * as AWSMOCK from 'aws-sdk-mock';
 import { DocumentClient } from 'aws-sdk/lib/dynamodb/document_client';
 import { expect } from 'chai';
+import PutItemInput = DocumentClient.PutItemInput;
+import GetItemInput = DocumentClient.GetItemInput;
 import * as lambdaLocal from 'lambda-local';
 import { it } from 'mocha';
 import * as nock from 'nock';
@@ -20,8 +23,6 @@ import { QuestionMarkValidator } from './QuestionMarkValidator';
 import { SessionAttributeValidator } from './SessionAttributeValidator';
 import { SpeechValidator } from './SpeechValidator';
 import { VideoAppValidator } from './VideoAppValidator';
-import PutItemInput = DocumentClient.PutItemInput;
-import GetItemInput = DocumentClient.GetItemInput;
 
 lambdaLocal.getLogger().level = 'error';
 
@@ -49,7 +50,7 @@ const upsServiceMock : { getProfileName : () => string; getProfileGivenName : ()
 
 export class AlexaTest {
 
-    private readonly handler : Function;
+    private readonly handler : Function | string;
     private settings : SkillSettings;
     private validators : ResponseValidator[];
 
@@ -57,7 +58,7 @@ export class AlexaTest {
     private partitionKeyName : string;
     private attributesName : string;
 
-    constructor(handler : Function, settings : SkillSettings) {
+    constructor(handler : Function | string, settings : SkillSettings) {
         this.handler = handler;
         this.settings = settings;
         this.validators = [
@@ -141,22 +142,19 @@ export class AlexaTest {
                 request.context.System.user.accessToken = currentItem.withUserAccessToken;
             }
 
-            this.mockDynamoDB(settings, currentItem);
+            let invoker : any;
+            if (typeof settings.handler === 'function') {
+                invoker = this.invokeFunction(settings, currentItem, request);
+            } else if (settings.handler.startsWith('lambda:')) {
+                invoker = this.invokeLambda(settings.handler.substr(7), settings, currentItem, request);
+            } else if (settings.handler.startsWith('cfn:')) {
+                const cfnParts = settings.handler.split(':');
+                invoker = this.invokeLambdaCloudformation(cfnParts[1], cfnParts[2], settings, currentItem, request);
+            } else {
+                throw 'Invalid handler configuration';
+            }
 
-            const interceptors = this.mockProfileAPI(currentItem);
-
-            lambdaLocal.execute({
-                event: request,
-                lambdaFunc: settings,
-                lambdaHandler: 'handler',
-                timeoutMs: 5000,
-            }).then((response) => {
-                if (response.toJSON) {
-                    response = response.toJSON();
-                }
-
-                interceptors.forEach((value) => nock.removeInterceptor(value));
-
+            invoker.then((response) => {
                 if (settings.debug) {
                     console.log(JSON.stringify(response, null, 2));
                 }
@@ -168,10 +166,64 @@ export class AlexaTest {
                 if (currentItem.callback) {
                     currentItem.callback(response);
                 }
-
                 this.runSingleTest(settings, sequence, sequenceIndex + 1, response.sessionAttributes, done);
             }).catch(done);
         }
+    }
+
+    private async invokeLambdaCloudformation(stackName : string, logicalName : string, settings : TestSettings, currentItem : SequenceItem, request : RequestEnvelope) : Promise<any> {
+        const cfnRes = await new CFNClient().describeStackResource({
+            StackName: stackName,
+            LogicalResourceId: logicalName,
+        }).promise();
+        if (cfnRes.StackResourceDetail.ResourceType !== 'AWS::Lambda::Function') {
+            throw 'CloudFormation resource must be of type AWS::Lambda::Function';
+        }
+        const lambdaName = cfnRes.StackResourceDetail.PhysicalResourceId;
+        return this.invokeLambda(lambdaName, settings, currentItem, request);
+    }
+
+    private async invokeLambda(name : string, settings : TestSettings, currentItem : SequenceItem, request : RequestEnvelope) : Promise<any> {
+        if (this.containsMockSettings(currentItem)) {
+            throw 'Invalid test configuration found. Cannot mock DynamoDB or API calls when invoking remotely.';
+        }
+        const res = await new LambdaClient().invoke({
+            FunctionName: name,
+            InvocationType: 'RequestResponse',
+            LogType: 'None',
+            Payload: JSON.stringify(request),
+        }).promise();
+
+        const parsedResponse = JSON.parse(res.Payload.toString());
+        if (res.FunctionError) {
+            console.log(JSON.stringify(parsedResponse, null, 2));
+        }
+        return parsedResponse;
+    }
+
+    private containsMockSettings(currentItem : SequenceItem) : boolean {
+        return currentItem.hasOwnProperty('storesAttributes') ||
+        currentItem.hasOwnProperty('withProfile') ||
+        currentItem.hasOwnProperty('withStoredAttributes');
+    }
+
+    private invokeFunction(settings : TestSettings, currentItem : SequenceItem, request : RequestEnvelope) : Promise<any> {
+        this.mockDynamoDB(settings, currentItem);
+
+        const interceptors = this.mockProfileAPI(currentItem);
+
+        return lambdaLocal.execute({
+            event: request,
+            lambdaFunc: settings,
+            lambdaHandler: 'handler',
+            timeoutMs: 5000,
+        }).then((response) => {
+            if (response.toJSON) {
+                response = response.toJSON();
+            }
+            interceptors.forEach((value) => nock.removeInterceptor(value));
+            return response;
+        });
     }
 
     private mockDynamoDB(settings : TestSettings, currentItem : SequenceItem) : void {
@@ -206,7 +258,7 @@ export class AlexaTest {
                 const Item = {};
                 Item[this.partitionKeyName] = settings.userId;
                 Item[this.attributesName] = currentItem.withStoredAttributes || {};
-                callback(null, {TableName: this.dynamoDBTable, Item});
+                callback(null, { TableName: this.dynamoDBTable, Item });
             };
         }
     }
@@ -258,7 +310,7 @@ interface TestSettings {
     appId : string;
     userId : string;
     deviceId : string;
-    handler : Function;
+    handler : Function | string;
     sessionId : string;
     debug : boolean;